+++ examples/update-relayer/index.js
@@ -1,10 +1,10 @@
 require('dotenv').config();
 
-const { Platform } = require('@openzeppelin/platform-sdk');
+const { Defender } = require('@openzeppelin/defender-sdk');
 
 async function main() {
   const creds = { apiKey: process.env.API_KEY, apiSecret: process.env.API_SECRET };
-  const client = new Platform(creds);
+  const client = new Defender(creds);
 
   const currentRelayer = await client.relay.get({ relayerId: 'e65451f5-0f76-4961-8d29-33504319c5f8' });