+++ packages/relay/src/api/index.ts
@@ -1,4 +1,4 @@
-import { BaseApiClient } from '@openzeppelin/platform-sdk-base-client';
+import { BaseApiClient } from '@openzeppelin/defender-sdk-base-client';
 import {
   RelayerGetResponse,
   CreateRelayerRequest,
@@ -11,15 +11,15 @@ import {
 
 export class RelayClient extends BaseApiClient {
   protected getPoolId(): string {
-    return process.env.PLATFORM_POOL_ID || 'us-west-2_94f3puJWv';
+    return process.env.DEFENDER_POOL_ID || 'us-west-2_94f3puJWv';
   }
 
   protected getPoolClientId(): string {
-    return process.env.PLATFORM_POOL_CLIENT_ID || '40e58hbc7pktmnp9i26hh5nsav';
+    return process.env.DEFENDER_POOL_CLIENT_ID || '40e58hbc7pktmnp9i26hh5nsav';
   }
 
   protected getApiUrl(): string {
-    return process.env.PLATFORM_API_URL || 'https://defender-api.openzeppelin.com/relayer/';
+    return process.env.DEFENDER_API_URL || 'https://platform-api.openzeppelin.com/relayer/';
   }
 
   public async get({ relayerId }: { relayerId: string }): Promise<RelayerGetResponse> {