+++ examples/create-relayer-key/CHANGELOG.md
@@ -23,7 +23,7 @@
 ### Patch Changes
 
 - Updated dependencies [efefd68]
-  - @openzeppelin/platform-sdk@0.3.0
+  - @openzeppelin/defender-sdk@0.3.0
 
 ## 0.2.0
 
@@ -42,4 +42,4 @@
 ### Patch Changes
 
 - Updated dependencies [496f88d]
-  - @openzeppelin/platform-sdk@0.2.0
+  - @openzeppelin/defender-sdk@0.2.0