+++ examples/relayer-signer-actions/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @openzeppelin/platform-sdk-example-relay-signer-action
+# @openzeppelin/defender-sdk-example-relay-signer-action
 
 ## 0.3.0
 
@@ -16,7 +16,7 @@
 ### Patch Changes
 
 - Updated dependencies [efefd68]
-  - @openzeppelin/platform-sdk@0.3.0
+  - @openzeppelin/defender-sdk@0.3.0
 
 ## 0.2.0
 
@@ -35,4 +35,4 @@
 ### Patch Changes
 
 - Updated dependencies [496f88d]
-  - @openzeppelin/platform-sdk@0.2.0
+  - @openzeppelin/defender-sdk@0.2.0