+++ examples/update-relayer/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @openzeppelin/platform-sdk-example-update-relayer
+# @openzeppelin/defender-sdk-example-update-relayer
 
 ## 0.3.0
 
@@ -23,7 +23,7 @@
 ### Patch Changes
 
 - Updated dependencies [efefd68]
-  - @openzeppelin/platform-sdk@0.3.0
+  - @openzeppelin/defender-sdk@0.3.0
 
 ## 0.2.0
 
@@ -42,4 +42,4 @@
 ### Patch Changes
 
 - Updated dependencies [496f88d]
-  - @openzeppelin/platform-sdk@0.2.0
+  - @openzeppelin/defender-sdk@0.2.0