+++ packages/base/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @openzeppelin/platform-sdk-base-client
+# @openzeppelin/defender-sdk-base-client
 
 ## 0.4.0
 
@@ -10,7 +10,7 @@
 
 ### Minor Changes
 
-- efefd68: feat: Add support of instantiating platform-sdk package from Platform Actions
+- efefd68: feat: Add support of instantiating defender-sdk package from Platform Actions
 
 ## 0.2.0